requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose[cryptography]
pytest
passlib
httpx
python-multipart
pydantic-settings